[fix] danmu title match
