+++ package.json
@@ -4,7 +4,7 @@
     "description": "A fntv app built with Electron",
     "main": "dest/main/main.js",
     "scripts": {
-        "start": "tsc && electron .",
+        "start": "npm run clean && tsc && electron .",
         "compile": "tsc",
         "watch": "tsc --watch",
         "clean": "rimraf dest",
@@ -23,6 +23,7 @@
         "@types/http-cache-semantics": "^4.0.4",
         "@types/keyv": "^3.1.4",
         "@types/ms": "^2.1.0",
+        "@types/multistream": "^4.1.3",
         "@types/node": "^20.19.11",
         "@types/plist": "^3.0.5",
         "@types/responselike": "^1.0.3",
@@ -41,6 +42,8 @@
         "axios": "^1.11.0",
         "express": "^5.1.0",
         "http-proxy-middleware": "^3.0.5",
+        "http-range-fetcher": "^2.0.0",
+        "multistream": "^4.1.0",
         "node-cache": "^5.1.2",
         "node-mpv-2": "^2.0.0-beta.3",
         "semver": "^7.7.2",
